// packages/bridges/src/model.generated.ts
export const NATIVE_ROUTES = [
  {
    id: "mezo-native-usdc-ethereum-to-mezo",
    direction: "inbound-to-mezo",
    source: {
      networkId: "ethereum-mainnet",
      contractId: "bridge.native-mezo-bridge",
      token: "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48",
    },
    destination: {
      networkId: "mezo-mainnet",
      contractId: "bridge.native-assets-precompile",
      token: "0x04671c72aab5ac02a03c1098314b1bb6b560c197",
    },
    targetChain: null,
  },
  {
    id: "mezo-native-btc-mezo-to-ethereum",
    direction: "outbound-from-mezo",
    source: {
      networkId: "mezo-mainnet",
      contractId: "bridge.native-assets-precompile",
      token: "0x7b7c000000000000000000000000000000000000",
    },
    destination: {
      networkId: "ethereum-mainnet",
      contractId: "bridge.native-mezo-bridge",
      token: "0x18084fba666a33d37592fa2633fd49a74dd93a88",
    },
    targetChain: 0,
  },
] as const;